+++ Code/Mixduino/src/touchbar.hpp
@@ -8,7 +8,7 @@ class TouchBar {
 public:
   TouchBar(uint16_t i2cAddr);
   void begin();
-  void read(EventManager &em, uint8_t key);
+  void read(EventManager &em, uint16_t key);
 private:
   Trill* m_trill{nullptr};
   const uint16_t m_i2cAddr{0};
